requirements.txt
websocket-client
python-dotenv